requirements.txt
absl-py==2.1.0
astunparse==1.6.3
certifi==2024.2.2
flatbuffers==23.5.26
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.11.0
keras==2.15.0
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
ml-dtypes==0.2.0
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
Pillow==10.2.0
protobuf==4.25.3
requests==2.31.0
six==1.16.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing-extensions==4.10.0
Werkzeug==3.0.2
wrapt==1.14.1
# ... and maybe a few others
